cmake: Add INSTALL_NAME_DIR to library to fix OS X 10.11 dynamic path issues
OS X 10.11 with System Integrity Protection (SIP) [1] turned on (which is
the default), purges all DYLD_xxx env. var's before executing a program.
This means that dynamic libraries can not be found via DYLD_xxx paths, and
so all dynamic library dependencies must have a full path specified.

Without this property set, dependent libraries/executables that dynamically
link with log4cpp end up without the path information, and thus fail to
load in 10.11 w/ SIP enabled. Even if CMake provides a full path to the
log4cpp library, the output dependent dynamic library has a log4cpp
entry like

    liblog4cpp.6.0.dylib (compatibility version 6.0.0, current version 6.0.0)

but with this property set in log4cpp (and no changes in the dependent
package), then the resulting log4cpp entry in the dependent library is

    /opt/orocos/lib/liblog4cpp.6.0.dylib (compatibility version 6.0.0, current version 6.0.0)

Not entirely sure why this entry fixes the problem, but it does.

[1] https://support.apple.com/en-us/HT204899