cmake: Force use of 64-bit file offsets
Copes with the change in SHA 493212f, which reverted the hard coded
version of this built into the code.

RDT requires >2GB files even on 32-bit filesystems, and thus this
is required. We don't want to enable this in the parent's do-build.sh
script, as that may have unintended affects on all the other ROS and
Orocos code. Instead we chose to continue to hard code this here, but
use a more portable manner.